refactor(scraper): use full brand name for scripts
This commit renames the `marant.ts` file to be `isabel-marant.ts` as
I'm now using the full brand name for scraping scripts.
