+++ scraper/src/main.ts
@@ -1,3 +1,3 @@
-import { scrape } from './marant';
+import { scrape } from './isabel-marant';
 
 void scrape();
